[INJIMOB-3598] refactor: modify folder structure of response
