+++ Sources/VCIClient/authorizationCodeFlow/InteractiveAuthorization/Request/AuthorizationDetail.swift
@@ -1,0 +1,19 @@
+import Foundation
+
+struct AuthorizationDetail: Codable {
+    let type: String
+    let credentialConfigurationId: [String]
+    let claims: [String: String]?
+
+    enum CodingKeys: String, CodingKey {
+        case type
+        case credentialConfigurationId = "credential_configuration_id"
+        case claims
+    }
+
+    init(type: String, credentialConfigurationId: [String], claims: [String: String]? = nil) {
+        self.type = type
+        self.credentialConfigurationId = credentialConfigurationId
+        self.claims = claims
+    }
+}
