+++ Sources/VCIClient/authorizationCodeFlow/InteractiveAuthorization/Request/AuthorizationRequestData.swift
